+++ terraform/main.tf
@@ -2,7 +2,7 @@
 // See Also: shared-main.tf
 // -----------------------------------------------------------------------------
 locals {
-  private_tier_ids = [for id in data.aws_subnet_ids.private.ids : id]
+  private_tier_ids = [for id in data.aws_subnets.private.ids : id]
 }
 
 data "aws_acm_certificate" "cengage_info" {
@@ -12,15 +12,21 @@ data "aws_acm_certificate" "cengage_info" {
 
 // -- Load Necessary Subnet Ids --
 // private/app tier subnets
-data "aws_subnet_ids" "private" {
-  vpc_id = var.vpc_id
-  tags   = { Name = "*-${local.env_type}-app-tier*" }
+data "aws_subnets" "private" {
+  filter {
+    name   = "vpc-id"
+    values = [var.vpc_id]
+  }
+  tags = { Name = "*-${local.env_type}-app-tier*" }
 }
 
 // public/web tier subnets
-data "aws_subnet_ids" "public" {
-  vpc_id = var.vpc_id
-  tags   = { Name = "*-${local.env_type}-web-tier*" }
+data "aws_subnets" "public" {
+  filter {
+    name   = "vpc-id"
+    values = [var.vpc_id]
+  }
+  tags = { Name = "*-${local.env_type}-web-tier*" }
 }
 
 data "aws_ami" "base_image" {
@@ -38,7 +44,7 @@ data "aws_ami" "base_image" {
 
 # Instance Security Group
 module "sec_grp_instance" {
-  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.2.0"
+  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.3.0"
   name        = "${var.app_name}-${var.app_env}-instance"
   description = "${var.app_name}-${var.app_env} instance security group"
   vpc_id      = var.vpc_id
@@ -107,7 +113,7 @@ resource "aws_cloudwatch_metric_alarm" "reboot_on_fail" {
 
 # ALB Security Group - Public
 module "sec_grp_alb_public" {
-  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.2.0"
+  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.3.0"
   name        = "${var.app_name}-${var.app_env}-alb-public"
   description = "${var.app_name}-${var.app_env} Public ALB security group"
   vpc_id      = var.vpc_id
@@ -119,7 +125,7 @@ module "sec_grp_alb_public" {
 
 # ALB Security Group - Private
 module "sec_grp_alb_private" {
-  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.2.0"
+  source      = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-sg-extended.git?ref=1.3.0"
   name        = "${var.app_name}-${var.app_env}-alb-private"
   description = "${var.app_name}-${var.app_env} Private ALB security group"
   vpc_id      = var.vpc_id
@@ -131,13 +137,13 @@ module "sec_grp_alb_private" {
 
 # Private ALB
 module "app_lb_private" {
-  source = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-alb.git?ref=1.3.0"
+  source = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-alb.git?ref=1.4.0"
 
   load_balancer_type = "application"
   name               = "${var.app_name}-${var.app_env}-private"
   internal           = true
   vpc_id             = var.vpc_id
-  subnets            = data.aws_subnet_ids.private.ids
+  subnets            = data.aws_subnets.private.ids
   security_groups    = [module.sec_grp_alb_private.this_security_group_id]
 
   tags    = local.common_tags
@@ -227,13 +233,13 @@ resource "aws_lb_target_group_attachment" "register_instances_private" {
 
 # Public ALB
 module "app_lb_public" {
-  source = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-alb.git?ref=1.3.0"
+  source = "git::ssh://git@stash.cengage.com:7999/tm/terraform-aws-alb.git?ref=1.4.0"
 
   load_balancer_type = "application"
   name               = "${var.app_name}-${var.app_env}-public"
   internal           = false
   vpc_id             = var.vpc_id
-  subnets            = data.aws_subnet_ids.public.ids
+  subnets            = data.aws_subnets.public.ids
   security_groups    = [module.sec_grp_alb_public.this_security_group_id]
 
   tags    = local.common_tags